Added missing constructor member var inits for some low-level classes and float vs. double cast/literal changes to quiet MSVS 2019 warnings

+++ src/inlines.h
@@ -372,11 +372,11 @@ inline bool line_poly_intersect(vector3d const &v1, point const &p1, point const
 
 inline bool point_in_ellipse(point const &p, point const &center, float rx, float ry) {
 	float const dx(center.x - p.x), dy(center.y - p.y);
-	return (dx*dx/(rx*rx) + dy*dy/(ry*ry) <= 1.0);
+	return (double(dx*dx/(rx*rx)) + double(dy*dy/(ry*ry)) <= 1.0);
 }
 inline bool point_in_ellipse_risq(point const &p, point const &center, float rx_inv_sq, float ry_inv_sq) {
 	float const dx(center.x - p.x), dy(center.y - p.y);
-	return (dx*dx*rx_inv_sq + dy*dy*ry_inv_sq <= 1.0);
+	return (double(dx*dx*rx_inv_sq) + double(dy*dy*ry_inv_sq) <= 1.0);
 }
